+++ example/lib/main.dart
@@ -3,6 +3,7 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:local_image_provider/device_image.dart';
 import 'package:local_image_provider/local_album.dart';
 import 'package:local_image_provider/local_image.dart';
 import 'package:local_image_provider/local_image_provider.dart';
@@ -18,12 +19,14 @@ class _MyAppState extends State<MyApp> {
   LocalImageProvider localImageProvider = LocalImageProvider();
   int _localImageCount = 0;
   bool _hasPermission = false;
+  bool _inStress = false;
   List<LocalImage> _localImages = [];
   List<LocalAlbum> _localAlbums = [];
-  Uint8List _imgBytes;
   bool _hasImage = false;
   String _imgSource;
   String _selectedId;
+  String _imgHeading = "most recent 100";
+  LocalImage _selectedImg;
 
   @override
   void initState() {
@@ -41,7 +44,7 @@ class _MyAppState extends State<MyApp> {
     try {
       hasPermission = await localImageProvider.initialize();
       if (hasPermission) {
-        localImages = await localImageProvider.findLatest(2);
+        localImages = await localImageProvider.findLatest(50);
         localAlbums = await localImageProvider.findAlbums(LocalAlbumType.all);
       }
     } on PlatformException catch (e) {
@@ -62,24 +65,64 @@ class _MyAppState extends State<MyApp> {
     });
   }
 
-  void switchAlbum( LocalAlbum album ) async {
-    List<LocalImage> albumImages = await localImageProvider.findImagesInAlbum(album.id, 100 );
+  void switchAlbum(LocalAlbum album) async {
+    List<LocalImage> albumImages =
+        await localImageProvider.findImagesInAlbum(album.id, 100);
     setState(() {
+      _imgHeading = "from album: ${album.title}";
       _localImages.clear();
       _localImages.addAll(albumImages);
     });
-    switchImage(album.coverImgId, 'Album');
+    switchImage(album.coverImg, 'Album');
   }
 
-  void switchImage(String imageId, String src) {
-    localImageProvider.imageBytes(imageId, 500, 500).then((img) {
+  void switchImage( LocalImage image, String src) {
       setState(() {
-        _imgBytes = img;
         _hasImage = true;
         _imgSource = src;
-        _selectedId = imageId;
-      });
+        _selectedImg = image;
     });
+  }
+
+  void stressTest() async {
+    if (_inStress) {
+      return;
+    }
+    print("Starting stress test");
+    _inStress = false;
+    int totalImagesLoaded = 0;
+    int totalAlbums = 0;
+    int totalBytesLoaded = 0;
+    DateTime start = DateTime.now();
+    for (int albumLoop = 0; albumLoop < 1000; ++albumLoop) {
+      List<LocalAlbum> localAlbums =
+          await localImageProvider.findAlbums(LocalAlbumType.all);
+      totalAlbums += localAlbums.length;
+      for (var album in localAlbums) {
+        Uint8List imgBytes =
+            await localImageProvider.imageBytes(album.coverImgId, 500, 500);
+        totalBytesLoaded += imgBytes.length;
+        ++totalImagesLoaded;
+        List<LocalImage> albumImages =
+            await localImageProvider.findImagesInAlbum(album.id, 100);
+        for (var albumImage in albumImages) {
+          Uint8List imgBytes =
+              await localImageProvider.imageBytes(albumImage.id, 500, 500);
+          totalBytesLoaded += imgBytes.length;
+          ++totalImagesLoaded;
+        }
+      }
+      DateTime current = DateTime.now();
+      int millisSoFar =
+          current.millisecondsSinceEpoch - start.millisecondsSinceEpoch;
+      double avgLoopTime = millisSoFar / (albumLoop + 1);
+      int secondsRemaining =
+          (avgLoopTime * (1000 - albumLoop - 1) / 1000).truncate();
+      print(
+          "$albumLoop, albums: $totalAlbums, images: $totalImagesLoaded, bytes: $totalBytesLoaded, remaining: $secondsRemaining");
+    }
+    _inStress = false;
+    print("Stress test complete");
   }
 
   @override
@@ -98,13 +141,36 @@ class _MyAppState extends State<MyApp> {
                     style: TextStyle(fontSize: 24),
                   ),
                   Divider(),
-                  Text('Images', style: TextStyle(fontSize: 20)),
+                  Flexible(
+                    child: FlatButton(
+                      child: Text('Stress Test'),
+                      onPressed: stressTest,
+                    ),
+                  ),
+                  Text('Albums', style: TextStyle(fontSize: 20)),
+                  Expanded(
+                    child: ListView(
+                      children: _localAlbums
+                          .map(
+                            (album) => GestureDetector(
+                                onTap: () => switchAlbum(album),
+                                child: Container(
+                                  padding: EdgeInsets.symmetric(vertical: 5),
+                                  child: Text(
+                                      'Title: ${album.title}; images: ${album.imageCount}, id: ${album.id}; cover Id: ${album.coverImg}'),
+                                )),
+                          )
+                          .toList(),
+                    ),
+                  ),
+                  Divider(),
+                  Text('Images - $_imgHeading', style: TextStyle(fontSize: 20)),
                   Expanded(
                     child: ListView(
                       children: _localImages
                           .map(
                             (img) => GestureDetector(
-                              onTap: () => switchImage(img.id, "Image"),
+                              onTap: () => switchImage(img, 'Images'),
                               child: Container(
                                 padding: EdgeInsets.symmetric(vertical: 5),
                                 child: Text(
@@ -116,31 +182,14 @@ class _MyAppState extends State<MyApp> {
                     ),
                   ),
                   Divider(),
-                  Text('Albums', style: TextStyle(fontSize: 20)),
-                  Expanded(
-                    child: ListView(
-                      children: _localAlbums
-                          .map(
-                            (album) => GestureDetector(
-                                onTap: () =>
-                                    switchAlbum(album),
-                                child: Container(
-                                  padding: EdgeInsets.symmetric(vertical: 5),
-                                  child: Text(
-                                      'Title: ${album.title}; id: ${album.id}; cover Id: ${album.coverImgId}'),
-                                )),
-                          )
-                          .toList(),
-                    ),
-                  ),
                   Expanded(
                     child: _hasImage
                         ? Column(
                             children: <Widget>[
                               Text('Selected: $_imgSource'),
-                              Text('Image id: $_selectedId'),
+                              Text('Image id: ${_selectedImg.id}'),
                               Expanded(
-                                child: Image.memory(_imgBytes),
+                                child: Image( image: DeviceImage( _selectedImg )),
                               ),
                             ],
                           )